+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 - Fix remote state not being collected sometimes (leaving unkilled projectiles)
 - Fix two setInteval run loops running when using crossnode
+- Fix socket memory leaks after server close
 
 ## [0.5.6] 2025-10-31